Correct the implementation of domain models to allow domain events on Commit

+++ test/Neutralize.EFCore.Tests/Dapper/ITodoRepository.cs
@@ -2,7 +2,7 @@ using Neutralize.Repositories;
 
 namespace Neutralize.EFCore.Tests.Dapper
 {
-    public interface ITodoRepository : IRepository<ToDo, int>
+    public interface ITodoRepository : IRepository<ToDo, long>
     {
     }
 }
