+++ test/Neutralize.EFCore.Tests/Dapper/TodoEfCoreRepository.cs
@@ -1,6 +1,6 @@
 namespace Neutralize.EFCore.Tests.Dapper
 {
-    public class TodoEfCoreRepository : EfCoreRepository<EfCoreDbContext, ToDo, int>, ITodoRepository
+    public class TodoEfCoreRepository : EfCoreRepository<EfCoreDbContext, ToDo, long>, ITodoRepository
     {
         public TodoEfCoreRepository(EfCoreDbContext context) : base(context)
         {
